fix 1.4 version of sensu_ssl_tool.tar
A tragic search-and-replace accident in commit c908a257 modified
the X509v3 extentsion OIDs in this version of the ssl tool tarball.
As a result, certificates generated by the modified ssl tool do not
have the required "Web TLS Client Authentication" extension, and
therefore do not work for deploying Sensu + RabbitMQ with peer verification.

Oops!

This change restores the required extensions to certificates generated by the
ssl tool.

Signed-off-by: Cameron Johnston <cameron@sensu.io>