Add notice for build-essential prerequisite (#592)
As the 5-minute guide suggests installation on a brand-new system, minimal installations of the supported OSes will not include build-essential tools, and the mail handler given in the example will fail to install. It was easy enough to figure out from the output, but a notice would have made my installation smoother.