Update client-to-server transport documentation (#1123)
* Update Client-to-server transport section

The etcd-cert-file and etcd-key-file values were referencing the client.pem and client-key.pem. However, this will result in issues setting up a cluster and you'll likely receive errors indicating there are no IP SANs since the client cert contains zero SANs. However, using the cert and key that were generated for the specific backend will allow the backend nodes to connect over TLS.

* Update client-to-server transport yaml example

The example yaml config uses the client.pem and client-key.pem, but that does not work. This needs to be the backend cert and key for the specific backend you are setting up. If you use the client cert and key you're cluster will fail to setup.

* Updates for agents connecting to cluster

* Copy to 5.1