Add namespace field to EtcdReplicators
A last minute change, the EtcdReplicator datatype has gained a
Namespace field, in order to resolve some issues stemming from an
implementation detail regarding namespaces.

This field is not represented in the metadata, as it does not
pertain to the replicator itself, but rather the resources that the
replicator will be replicating.

Signed-off-by: Eric Chlebek <eric@sensu.io>